docker/api-processor/requirements.txt
numpy==1.26.4
fastapi==0.109.2
uvicorn==0.27.1
celery==5.3.6
redis==5.0.1
boto3==1.34.34
shapely==2.0.2
python-multipart==0.0.6
Pillow==10.2.0